split into seperate file and clean up header
