+++ include/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/05/07 09:59:02 by spyun         #+#    #+#                 */
-/*   Updated: 2025/05/13 10:49:17 by spyun         ########   odam.nl         */
+/*   Updated: 2025/05/13 12:01:10 by spyun         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,37 +69,45 @@ typedef struct s_game
 /* main.c */
 int				main(int argc, char **argv);
 
-/* validation.c */
-bool			check_map_extension(const char *filename);
-bool			check_texture_path(const char *path);
-bool			check_color_value(int r, int g, int b);
-int				check_map_chars(char **map);
+/* parser */
+/* check_map_borders.c */
+bool			check_map_borders(char **map, t_game *game);
 
 /* check_map_closure.c */
-bool	validate_map_closure(t_game *game);
+bool			is_walkable(char c);
+bool			validate_map_closure(t_game *game);
 
-/* parse.c */
-int				parse_texture_path(char *line, char **path);
-int				parse_color_value(char *line, t_color *color);
-int				parse_element(char *line, t_game *game);
-int				parse_file(char *filename, t_game *game);
+/* check_map_spaces.c */
+bool			check_spaces_surrounded(char **map, t_game *game);
 
 /* map_builder.c  */
 t_list			*create_map_node(char *line);
 int				add_first_line(t_list **map_lines, char *first_line);
 int				build_map_array(t_game *game, t_list *map_lines);
 
+/*  parse_file.c */
+int				parse_file(char *filename, t_game *game);
+
 /* parse_map.c */
 int				get_map_width(char **map);
 int				parse_map(int fd, t_game *game, char *line);
 
+/* parse.c */
+int				parse_element(char *line, t_game *game);
+
+/* validation.c */
+bool			check_map_extension(const char *filename);
+bool			check_texture_path(const char *path);
+bool			check_color_value(int r, int g, int b);
+int				check_map_chars(char **map);
+
 /* render*/
 void			render_minimap(t_game *game);
 
-/* error*/
+/* error.c*/
 void			ft_mlx_error(t_game *game);
 
-/* free_utils */
+/* free_utils.c */
 void			free_map(char **map, t_game *game);
 void			free_game(t_game *game);
