+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/05/07 09:52:59 by spyun         #+#    #+#                 */
-/*   Updated: 2025/05/12 16:36:25 by seungah       ########   odam.nl         */
+/*   Updated: 2025/05/13 11:57:09 by spyun         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,18 +30,18 @@ static bool	setup_game(int argc, char **argv, t_game *game)
 {
 	if (argc != 2)
 	{
-		ft_putendl_fd("Error\nUsage: ./cub3d <map.cub>", 2);
+		ft_putendl_fd("Error: Usage: ./cub3d <map.cub>", 2);
 		return (false);
 	}
 	if (!check_map_extension(argv[1]))
 	{
-		ft_putendl_fd("Error\nInvalid file extension. Expected .cub", 2);
+		ft_putendl_fd("Error: nInvalid file extension. Expected .cub", 2);
 		return (false);
 	}
 	init_game_state(game);
 	if (parse_file(argv[1], game) == -1)
 	{
-		ft_putendl_fd("Error\nFailed to parse map", 2);
+		ft_putendl_fd("Error: Failed to parse map", 2);
 		free_game(game);
 		return (false);
 	}
